+++ sign_net/utils.py
@@ -25,7 +25,7 @@ class PETransform(object):
         if n <= self.pos_enc_dim:
             position_encoding = F.pad(position_encoding, (0, self.pos_enc_dim - n + 1), value=float('0'))
             
-        data.pos_enc = position_encoding.to(torch.float16)
+        data.pos_enc = position_encoding.float()
         return data
 
     def _position_encoding(self, data, norm=None):
